Isolate test_wms_getprint_atlas
